[FEATURE] implementation jwt token for authentication and authorization, also user scope

+++ repository/todolist_repository.go
@@ -6,14 +6,16 @@ import (
 	"errors"
 	"todolist/helper"
 	"todolist/model/domain"
+	"todolist/model/dto"
 )
 
 type TodolistRepository interface {
 	Insert(ctx context.Context, tx *sql.Tx, todolistDomain domain.TodolistDomain) domain.TodolistDomain
 	Update(ctx context.Context, tx *sql.Tx, todolistDomain domain.TodolistDomain) domain.TodolistDomain
-	Delete(ctx context.Context, tx *sql.Tx, id int)
-	FindAll(ctx context.Context, tx *sql.Tx) ([]domain.TodolistDomain, error)
-	FindById(ctx context.Context, tx *sql.Tx, id int) (domain.TodolistDomain, error)
+	Delete(ctx context.Context, tx *sql.Tx, id int, userId int)
+	FindAll(ctx context.Context, tx *sql.Tx, userId int) ([]domain.TodolistDomain, error)
+	FindById(ctx context.Context, tx *sql.Tx, id int, userId int) (domain.TodolistDomain, error)
+	CheckTodolist(ctx context.Context, tx *sql.Tx, request dto.CheckTodolistRequest, id int, userId int) (domain.TodolistDomain, error)
 }
 
 type TodolistRepositoryImpl struct {
@@ -25,34 +27,34 @@ func NewTodolistRepository(db *sql.DB) TodolistRepository {
 }
 
 func (repository *TodolistRepositoryImpl) Insert(ctx context.Context, tx *sql.Tx, todolistDomain domain.TodolistDomain) domain.TodolistDomain {
-	SQL := "INSERT INTO mst_todolist (title, description) VALUES ($1, $2) RETURNING id, isDone"
+	SQL := "INSERT INTO mst_todolist (title, description, user_id) VALUES ($1, $2, $3) RETURNING id, isDone"
 
-	err := tx.QueryRowContext(ctx, SQL, todolistDomain.Title, todolistDomain.Description).Scan(&todolistDomain.Id, &todolistDomain.IsDone)
+	err := tx.QueryRowContext(ctx, SQL, todolistDomain.Title, todolistDomain.Description, todolistDomain.UserId).Scan(&todolistDomain.Id, &todolistDomain.IsDone)
 	helper.PanicIfError(err)
 
 	return todolistDomain
 }
 
 func (repository *TodolistRepositoryImpl) Update(ctx context.Context, tx *sql.Tx, todolistDomain domain.TodolistDomain) domain.TodolistDomain {
-	SQL := "update mst_todolist set title = $1, description = $2, isDone = $3 where id = $4"
+	SQL := "update mst_todolist set title = $1, description = $2, isDone = $3 where id = $4 and user_id = $5"
 
-	_, err := tx.ExecContext(ctx, SQL, todolistDomain.Title, todolistDomain.Description, todolistDomain.IsDone, todolistDomain.Id)
+	_, err := tx.ExecContext(ctx, SQL, todolistDomain.Title, todolistDomain.Description, todolistDomain.IsDone, todolistDomain.Id, todolistDomain.UserId)
 	helper.PanicIfError(err)
 
 	return todolistDomain
 }
 
-func (repository *TodolistRepositoryImpl) Delete(ctx context.Context, tx *sql.Tx, id int) {
-	SQL := "delete from mst_todolist where id = $1"
+func (repository *TodolistRepositoryImpl) Delete(ctx context.Context, tx *sql.Tx, id int, userId int) {
+	SQL := "delete from mst_todolist where id = $1 and user_id = $2"
 
-	_, err := tx.ExecContext(ctx, SQL, id)
+	_, err := tx.ExecContext(ctx, SQL, id, userId)
 	helper.PanicIfError(err)
 }
 
-func (repository *TodolistRepositoryImpl) FindAll(ctx context.Context, tx *sql.Tx) ([]domain.TodolistDomain, error) {
-	SQL := "select id, title, description, isDone from mst_todolist"
+func (repository *TodolistRepositoryImpl) FindAll(ctx context.Context, tx *sql.Tx, userId int) ([]domain.TodolistDomain, error) {
+	SQL := "select id, title, description, isDone from mst_todolist where user_id = $1"
 
-	rows, err := tx.QueryContext(ctx, SQL)
+	rows, err := tx.QueryContext(ctx, SQL, userId)
 	helper.PanicIfError(err)
 
 	var todolist []domain.TodolistDomain
@@ -75,13 +77,28 @@ func (repository *TodolistRepositoryImpl) FindAll(ctx context.Context, tx *sql.T
 	return todolist, nil
 }
 
-func (repository *TodolistRepositoryImpl) FindById(ctx context.Context, tx *sql.Tx, id int) (domain.TodolistDomain, error) {
-	SQL := "SELECT id, title, description, isDone FROM mst_todolist WHERE id = $1"
+func (repository *TodolistRepositoryImpl) FindById(ctx context.Context, tx *sql.Tx, id int, userId int) (domain.TodolistDomain, error) {
+	SQL := "SELECT id, title, description, isDone FROM mst_todolist WHERE id = $1 AND user_id = $2"
 
 	todolist := domain.TodolistDomain{}
 
-	err := tx.QueryRowContext(ctx, SQL, id).Scan(&todolist.Id, &todolist.Title, &todolist.Description, &todolist.IsDone)
-	helper.PanicIfError(err)
+	err := tx.QueryRowContext(ctx, SQL, id, userId).Scan(&todolist.Id, &todolist.Title, &todolist.Description, &todolist.IsDone)
+	if err != nil {
+		return todolist, err
+	}
+
+	return todolist, nil
+}
+
+func (repository *TodolistRepositoryImpl) CheckTodolist(ctx context.Context, tx *sql.Tx, request dto.CheckTodolistRequest, id int, userId int) (domain.TodolistDomain, error) {
+	SQL := "update mst_todolist set isDone = $1 where id = $2 and user_id = $3 returning id, title, description, isDone"
+
+	todolist := domain.TodolistDomain{}
+
+	err := tx.QueryRowContext(ctx, SQL, request.IsDone, id, userId).Scan(&todolist.Id, &todolist.Title, &todolist.Description, &todolist.IsDone)
+	if err != nil {
+		return todolist, err
+	}
 
 	return todolist, nil
 }
